Use pre-built image for action
This will use a prebuilt image that's automatically built for this
repository for each tag by GCB.  This should speed up action execution a
bit.

The version in this repo will still run from the local code though, so
we can still fix bugs.  That configuration lives in the `action-nightly`
subdirectory, and can also be referenced by external projects.