Remove descriptiveness check
Remove descriptiveness check. PR reviewers can ask for more details on a
per PR basis, and requiring descriptions meet a minimum length leads to
PR authors copying and pasting repeated text into the description to
make the check pass. There are definitely valid times when a description
is not needed (the title is sufficient).

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>